refactor: move query options to own package
